+++ src/main/java/mods/belgabor/amtweaker/mods/amt/AMT.java
@@ -12,8 +12,6 @@ public class AMT {
         MineTweakerAPI.registerClass(IceMaker.class);
         MineTweakerAPI.registerClass(Slag.class);
         MineTweakerAPI.registerClass(Battery.class);
-        /*
-        TODO: TeaMaker
-         */
+        MineTweakerAPI.registerClass(TeaMaker.class);
     }
 }